Rename Lang\Annotation into Lang\Attribute namespace
