+++ demos/Demo/Aspect/FluentInterfaceAspect.php
@@ -14,7 +14,7 @@ namespace Demo\Aspect;
 
 use Go\Aop\Aspect;
 use Go\Aop\Intercept\MethodInvocation;
-use Go\Lang\Annotation\Around;
+use Go\Lang\Attribute\Around;
 
 /**
  * Fluent interface aspect provides an easy way to reuse "chain" interface for classes
