+++ demos/Demo/Aspect/LoggingAspect.php
@@ -14,7 +14,7 @@ namespace Demo\Aspect;
 
 use Go\Aop\Aspect;
 use Go\Aop\Intercept\MethodInvocation;
-use Go\Lang\Annotation\Before;
+use Go\Lang\Attribute\Before;
 
 /**
  * Logging aspect
