+++ demos/Demo/Aspect/PropertyInterceptorAspect.php
@@ -14,7 +14,7 @@ namespace Demo\Aspect;
 
 use Go\Aop\Aspect;
 use Go\Aop\Intercept\FieldAccess;
-use Go\Lang\Annotation\Around;
+use Go\Lang\Attribute\Around;
 
 /**
  * Property interceptor can intercept an access to the public and protected properties
