+++ src/Lang/Attribute/Around.php
@@ -10,14 +10,14 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Go\Lang\Annotation;
+namespace Go\Lang\Attribute;
 
 /**
- * Before advice annotation
+ * Around advice annotation
  *
  * @Annotation
  * @Target({ "METHOD", "PROPERTY" })
  */
-class Before extends BaseInterceptor
+class Around extends BaseInterceptor
 {
 }
