+++ src/Lang/Attribute/BaseAnnotation.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Go\Lang\Annotation;
+namespace Go\Lang\Attribute;
 
 use BadMethodCallException;
