+++ tests/Fixtures/project/src/Aspect/DoSomethingAspect.php
@@ -5,7 +5,7 @@ namespace Go\Tests\TestProject\Aspect;
 
 use Go\Aop\Aspect;
 use Go\Aop\Intercept\MethodInvocation;
-use Go\Lang\Annotation as Pointcut;
+use Go\Lang\Attribute as Pointcut;
 
 class DoSomethingAspect implements Aspect
 {
